CNN
This project involves the classification of skin cancer using a convolutional neural network (CNN) model. The goal is to develop a model that can accurately classify skin images into different categories corresponding to different types of skin cancer.

The project follows the typical machine learning workflow, including data preprocessing, model construction, training, and evaluation.

Here is a step-by-step description of the project:

Data Preparation: The project starts with collecting a dataset of skin images. The dataset contains images of different types of skin cancer. The images are organized into different classes based on the type of cancer. The dataset is split into a training set and a validation set.

Data Augmentation: To address the class imbalance issue and improve model performance, data augmentation is applied to the training dataset. Augmentor, a Python library, is used to generate additional training samples by applying random rotations, flips, and zooms to the images.

Model Construction: The CNN model is constructed using the Keras library. The model architecture consists of several convolutional layers with activation functions, max pooling layers, and a dropout layer for regularization. The model ends with a fully connected layer followed by a softmax activation function to output the probabilities of different classes.

Model Compilation: The model is compiled with an optimizer (Adam), a loss function (Sparse Categorical Crossentropy), and evaluation metrics (accuracy). This configuration specifies how the model will be trained and evaluated.

Model Training: The model is trained using the training dataset. The training process involves iterating over the dataset for a specified number of epochs. In each epoch, the model makes predictions on the training samples, compares them with the ground truth labels, and adjusts its parameters using backpropagation and gradient descent to minimize the loss.

Model Evaluation: After training, the model is evaluated using the validation dataset. The evaluation process involves feeding the validation images into the trained model and calculating the loss and accuracy metrics. The evaluation results provide insights into the model's performance on unseen data.

Results Analysis: The training and validation loss/accuracy curves are plotted using Matplotlib to visualize the model's learning progress and identify any signs of overfitting or underfitting. The curves help determine if the model has learned effectively or if adjustments are needed.

Throughout the project, various technologies like TensorFlow, Keras, NumPy, PIL, Matplotlib, and Augmentor are utilized to handle data, construct the model, preprocess images, and visualize results.

The ultimate goal of this project is to develop a reliable and accurate skin cancer classification model that can assist in early detection and diagnosis, contributing to improved patient outcomes and healthcare practices.